+++ packages/teamscale-coverage-collector/test/DataStorage.test.ts
@@ -4,7 +4,6 @@ import { StdConsoleLogger } from '../src/utils/StdConsoleLogger';
 import tmp from 'tmp';
 import * as dat from 'date-and-time';
 import path from 'path';
-import * as fs from 'fs';
 
 test('Test if coverage is aggregated in the storage', () => {
 	const storage = new ProjectCoverage('42');
@@ -12,6 +11,7 @@ test('Test if coverage is aggregated in the storage', () => {
 	storage.putLine('main.ts', 2);
 	storage.putLine('main.ts', 4);
 	const lines = new Set();
+
 	for (const coverage of storage.getCoverage()) {
 		coverage.coveredLines.forEach(l => lines.add(l));
 	}
@@ -21,68 +21,46 @@ test('Test if coverage is aggregated in the storage', () => {
 	expect(lines).not.toContain(3);
 });
 
-test('Test that coverage files are timestamped before file ending', () => {
+test('Test that timestamped coverage file are created inside the target directory', () => {
 	const dataStorage = new DataStorage(createLogger());
 	const targetDir = tmp.dirSync({ unsafeCleanup: true });
-	const targetFileName = path.join(targetDir.name, 'coverage.simple');
 	const date = new Date();
 
-	dataStorage.dumpToSimpleCoverageFile(targetFileName, date);
+	const [coverageFile] = dataStorage.dumpToSimpleCoverageFile(targetDir.name, date);
 
 	const formattedDate = dat.format(date, dataStorage.DATE_FORMAT);
-	const coverageFiles = fs.readdirSync(targetDir.name);
-	expect(coverageFiles.length).toBe(1);
-	expect(coverageFiles[0]).toBe(`coverage-${formattedDate}.simple`);
+	expect(path.dirname(coverageFile)).toBe(targetDir.name);
+	expect(coverageFile).toContain(`coverage-${formattedDate}.simple`);
 
 	targetDir.removeCallback();
 });
 
-test('Test that coverage files without file ending are timestamped', () => {
+test('Test that coverage folder paths are trimmed', () => {
 	const dataStorage = new DataStorage(createLogger());
-	const targetDir = tmp.dirSync({ unsafeCleanup: true });
-	const targetFileName = path.join(targetDir.name, 'coverage');
+	const targetDirWithoutLeadingAndTailingSpaces = tmp.dirSync({ unsafeCleanup: true });
+	const targetDirWithLeadingAndTailingSpaces = ` ${targetDirWithoutLeadingAndTailingSpaces.name} `;
 	const date = new Date();
 
-	dataStorage.dumpToSimpleCoverageFile(targetFileName, date);
+	const [coverageFile] = dataStorage.dumpToSimpleCoverageFile(targetDirWithLeadingAndTailingSpaces, date);
 
-	const formattedDate = dat.format(date, dataStorage.DATE_FORMAT);
-	const coverageFiles = fs.readdirSync(targetDir.name);
-	expect(coverageFiles.length).toBe(1);
-	expect(coverageFiles[0]).toBe(`coverage-${formattedDate}`);
+	expect(path.dirname(coverageFile)).toBe(targetDirWithoutLeadingAndTailingSpaces.name);
 
-	targetDir.removeCallback();
-});
-
-test('Test that coverage file paths are trimmed', () => {
-	const dataStorage = new DataStorage(createLogger());
-	const targetDir = tmp.dirSync({ unsafeCleanup: true });
-	const targetFilePathWithLeadingAndTailingSpaces = ` ${path.join(targetDir.name, 'coverage')} `;
-	const date = new Date();
-
-	dataStorage.dumpToSimpleCoverageFile(targetFilePathWithLeadingAndTailingSpaces, date);
-
-	const formattedDate = dat.format(date, dataStorage.DATE_FORMAT);
-	const coverageFiles = fs.readdirSync(targetDir.name);
-	expect(coverageFiles.length).toBe(1);
-	expect(coverageFiles[0]).toBe(`coverage-${formattedDate}`);
-
-	targetDir.removeCallback();
+	targetDirWithoutLeadingAndTailingSpaces.removeCallback();
 });
 
 test('Coverage is reset after dump', () => {
 	const dataStorage = new DataStorage(createLogger());
 	const targetDir = tmp.dirSync({ unsafeCleanup: true });
-	const targetFileName = path.join(targetDir.name, 'coverage');
 	const firstDumpDate = new Date();
 	const project = 'test_project';
 	const coveredFile = 'test_file';
 	const coveredLines = [1, 2, 3];
 
 	dataStorage.putCoverage(project, coveredFile, coveredLines);
-	const [, firstWrittenLines] = dataStorage.dumpToSimpleCoverageFile(targetFileName, firstDumpDate);
+	const [, firstWrittenLines] = dataStorage.dumpToSimpleCoverageFile(targetDir.name, firstDumpDate);
 
 	const secondDumpDate = new Date();
-	const [, secondWrittenLines] = dataStorage.dumpToSimpleCoverageFile(targetFileName, secondDumpDate);
+	const [, secondWrittenLines] = dataStorage.dumpToSimpleCoverageFile(targetDir.name, secondDumpDate);
 
 	expect(firstWrittenLines).toBeGreaterThan(0);
 	expect(secondWrittenLines).toBe(0);