+++ packages/teamscale-coverage-collector/CHANGELOG.md
@@ -1,7 +1,9 @@
 We use [Semantic Versioning](https://semver.org/).
 
 # New Release
-- [fix] Reset coverage after dump and add timestamps to coverage files
+- [fix] Reset coverage after dump
+- [feature] Coverage files are now timestamped.
+- [deprecation] Deprecated `dump_to_file` in favor of `dump_to_folder`. If `dump_to_file` is still used, a folder will be created at the provided path and timestamped coverage files will be saved within it.
 
 # 0.0.1-beta.36